+++ aDVanceERP.Modulos.Inventario/MVP/Vistas/Producto/Plantillas/IVistaRegistroProducto.cs
@@ -20,7 +20,7 @@ public interface IVistaRegistroProducto : IVistaRegistro {
     decimal PrecioCompraBase { get; set; }
     decimal PrecioVentaBase { get; set; }
     string? NombreAlmacen { get; set; }
-    float StockInicial { get; set; }
+    decimal StockInicial { get; set; }
 
     event EventHandler? RegistrarUnidadMedida;
     event EventHandler? RegistrarTipoMateriaPrima;
